requirements.txt
crewai>=0.130.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
pypdf>=4.2.0
tavily-python>=0.5.0
google-api-python-client>=2.140.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
streamlit>=1.37.0
beautifulsoup4>=4.12.3
openai>=1.83.0
pdfplumber>=0.11.4
streamlit-autorefresh>=1.0.1
